==> scripts/00-simulation.R <==
simulated_data <-
  tibble(
    'Background_characteristic' = c(
      "Urban",
      "Rural",
      "Java-Bali",
      "Outer Java-Bali I",
      "outer Java-Bali II",
      "Jakarta",
      "West Java",
      "Central Java",
      "Yogyakarta",
      "East Java",
      "Bali",
      "None Education",
      "Some primary",
      "Primary completed",
      "Secondary or more",
      "Total"
    ),
    'no_children' = runif(n = 16,
                          min = 0, 
                          max = 100),
    'one_children' = runif(n = 16,
                          min = 0, 
                          max = 100),
    'two_children' = runif(n = 16,
                          min = 0, 
                          max = 100),
    'three_children' = runif(n = 16,
                          min = 0, 
                          max = 100),
    'four_children' = runif(n = 16,
                          min = 0, 
                          max = 100),
    'five_children' = runif(n = 16,
                          min = 0, 
                          max = 100),
    'six_and_more_children' = runif(n = 16,
                          min = 0, 
                          max = 100),
    'total' = runif(n = 16,
                          min = 0, 
                          max = 100),
    )
